// apps/sync-service/src/scripts/types.ts
export type Generated<T> =
  T extends ColumnType<infer S, infer I, infer U>
    ? ColumnType<S, I | undefined, U>
    : ColumnType<T, T | undefined, T>
// ...
export interface MasterMaterials {
  bpom_code: string | null
  code: string | null
  created_at: Date
  created_by: number | null
  deleted_at: Date | null
  deleted_by: number | null
  description: string | null
  id: Generated<number>
  is_addremove: Generated<number>
  is_openvial: Generated<number>
  is_so: Generated<number | null>
  is_stockcount: Generated<number>
  is_vaccine: number | null
  kfa_code: string | null
  kfa_level_id: number | null
  managed_in_batch: Generated<number>
  name: string | null
  need_sequence: number | null
  parent_id: number | null
  pieces_per_unit: Generated<number>
  range_temperature_id: number | null
  status: Generated<number>
  temperature_max: number | null
  temperature_min: number | null
  temperature_sensitive: number | null
  unit: string | null
  unit_of_distribution: string | null
  updated_at: Date
  updated_by: number | null
}
// ...
export interface Materials {
  bpom_code: number | null
  code: string | null
  created_at: Date
  created_by: number | null
  deleted_at: Date | null
  deleted_by: number | null
  description: string | null
  id: Generated<number>
  is_addremove: Generated<number>
  is_stockcount: Generated<number | null>
  is_vaccine: Generated<number>
  managed_in_batch: Generated<number>
  master_activity_id: number | null
  master_material_id: number | null
  name: string | null
  pieces_per_unit: Generated<number>
  status: Generated<number | null>
  temperature_max: number | null
  temperature_min: number | null
  temperature_sensitive: number | null
  unit: string | null
  updated_at: Date
  updated_by: number | null
}